# public/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Fleet Finance Manger",
        "description": "Fleet Finance Manger - API for Android and IOS Development",
        "termsOfService": "All rights reserved to Greefi Technologies",
        "contact": {
            "email": "hello@greefitech.com"
        },
        "version": "1.0.0"
    },
    "host": "localhost:9000",
    "basePath": "/",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/api/trip/create": {
            "post": {
                "tags": [
                    "Trip"
                ],
                "description": "Create New Trip",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "vehicleId",
                        "in": "query",
                        "description": "Enter Vehicle Id",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "tripName",
                        "in": "query",
                        "description": "Enter Trip Name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "dateFrom",
                        "in": "query",
                        "description": "Enter Date yyyy-mm-dd",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "dateTo",
                        "in": "query",
                        "description": "Enter Date to yyyy-mm-dd",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "startKm",
                        "in": "query",
                        "description": "Enter Start Km",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "endKm",
                        "in": "query",
                        "description": "Enter End Km",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "staff1",
                        "in": "query",
                        "description": "Enter Staff 1 ID",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "staff2",
                        "in": "query",
                        "description": "Enter Staff 2 ID",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "staff3",
                        "in": "query",
                        "description": "Enter Staff 3 ID",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "advance",
                        "in": "query",
                        "description": "Enter Advance Amount",
                        "required": false,
                        "type": "number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Invalid Data"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/trips": {
            "get": {
                "tags": [
                    "Trip"
                ],
                "description": "Return all Trips",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Invalid Data"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/trip/{id}/edit": {
            "get": {
                "tags": [
                    "Trip"
                ],
                "description": "Return particular Trip",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Invalid Data"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/trip/{id}/update": {
            "post": {
                "tags": [
                    "Trip"
                ],
                "description": "Update Trip",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "vehicleId",
                        "in": "query",
                        "description": "Enter Vehicle Id",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "tripName",
                        "in": "query",
                        "description": "Enter Trip Name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "dateFrom",
                        "in": "query",
                        "description": "Enter Date yyyy-mm-dd",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "dateTo",
                        "in": "query",
                        "description": "Enter Date to yyyy-mm-dd",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "startKm",
                        "in": "query",
                        "description": "Enter Start Km",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "endKm",
                        "in": "query",
                        "description": "Enter End Km",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "staff1",
                        "in": "query",
                        "description": "Enter Staff 1 ID",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "staff2",
                        "in": "query",
                        "description": "Enter Staff 2 ID",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "staff3",
                        "in": "query",
                        "description": "Enter Staff 3 ID",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "advance",
                        "in": "query",
                        "description": "Enter Advance Amount",
                        "required": false,
                        "type": "number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Invalid Data"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
...
}